Add test cases for truncated call vcf
There are two ways in which a variant can be no call. The variant can be
present in the vcf file with a ./. genotype, or the variant can be
missing from the vcf file altogether. This commit adds a test case to
verify that no call sites are counted properly when they are missing
from the called vcf file.

Sites that are missing from the positive vcf file do not count towards
alleles_no_call, since we do not care about variants which are not
present in the positive vcf file.