Compare allele genotypes directly
Previously we used cyvcf2 gt_types, which indicate whether a site is
hom_ref, heterozygous, hom_alt or unknown. A limitation of this approach
is that it only makes sense when the two vcf files have the same REF and
ALT alleles for a given position.

If this was not the case, vtools would silently skip the variant.
However, even when the ALT alleles are different between two vcf files,
the actual genotypes can still be compared. This was implemented in this
commit, which also fixes issue #8.