Add check for decomposed calls
Before, we would iterate over decomposed called variants until we found
the variant where the REF and ALT matched the one in the positive vcf.
However, this meant that sites where the called vcf had different ALTs
could not be processed.

To resolve this, we had to drop support for decomposed variants.