Fixed bug with partial no calls
To fix this, we had to rewrite the logic used when comparing variants.
As a side effect, 'allele_no_call' is now only counted for partial calls
when they occur in the call set. The test case
test_partial_positive_no_call was updated to reflect this change.

Also fixed an error where the known and called variants were switched
when they were being passed to parse_variants.